\name{X2way}
\alias{X2way}
\docType{data}
\title{Two way matrix}
\description{The data contains the scores of 8 persons on 10 personality traits}
\usage{data(X2way)}
\format{
Dispositions are the following:
    \cr{Fear to be refused}
    \cr{Kindness}
    \cr{Importance of others' judgments}
    \cr{Altruism}
    \cr{Neuroticism}
    \cr{Opennes}
    \cr{Being strict to oneself}
    \cr{Low selfesteem}
    \cr{Conscientiousness}
    \cr{Depression}
  }
\examples{
data(X2way)
str(X2way)
}
\keyword{datasets}
